+++ Assets/UI/TimerUI.cs
@@ -14,35 +14,70 @@ public class TimerUI : MonoBehaviour
     [SerializeField] private Ease _rotationEase;
 
     private float _elapsedTime;
+    private bool _isTimerRunning;
+    private Tween _rotationTween;
 
     void Start()
     {
         PlayRotationAnimation();
         StartTimer();
+
+        GameManager.Instance.OnGameStateChanged += OnGameStateChanged;
+    }
+
+    private void OnGameStateChanged(GameState gameState)
+    {
+        switch (gameState)
+        {
+            case GameState.Pause:
+                PauseTimer();
+                break;
+
+            case GameState.Resume:
+                ResumeTimer();
+                break;
+        }
     }
 
     private void PlayRotationAnimation()
     {
         Vector3 _rotateDirection = new Vector3(0f, 0f, -360f);
-        _timerRotatableTransform.DORotate(_rotateDirection, _timer, RotateMode.FastBeyond360)
+        _rotationTween = _timerRotatableTransform.DORotate(_rotateDirection, _timer, RotateMode.FastBeyond360)
             .SetLoops(-1, LoopType.Restart)
             .SetEase(_rotationEase);
     }
 
     private void StartTimer()
     {
+        _isTimerRunning = true;
         _elapsedTime = 0.0f;
         InvokeRepeating(nameof(UpdateTimerUI), 0f, 1f);
     }
 
     private void UpdateTimerUI()
     {
-
+        if (!_isTimerRunning)
+            return;
         _elapsedTime += 1f;
-
         int minutes = Mathf.FloorToInt(_elapsedTime / 60f);
         int seconds = Mathf.FloorToInt(_elapsedTime % 60f);
 
         _timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
+    }
+    private void PauseTimer()
+    {
+        _isTimerRunning = false;
+        CancelInvoke(nameof(UpdateTimerUI));
+        _rotationTween.Pause();
+    }
+     
+    private void ResumeTimer()
+    {
+        if (!_isTimerRunning)
+        {
+            _isTimerRunning = true;
+            InvokeRepeating(nameof(UpdateTimerUI), 0f, 1f);
+            _rotationTween.Play();
+        }
     }
 }