+++ Assets/_GameAssets/Scripts/Managers/GameManager.cs
@@ -4,7 +4,9 @@ using UnityEngine;
 public class GameManager : MonoBehaviour
 {
     public event Action<int> OnEggCollected;
+    public event Action<GameState> OnGameStateChanged;
     public static GameManager Instance { get; private set; }
+    private GameState _currentGameState;
 
     [SerializeField] private int maxEggCount = 5;
 
@@ -22,6 +24,18 @@ public class GameManager : MonoBehaviour
         DontDestroyOnLoad(gameObject); // Sahneler arasında yok olmasın (opsiyonel)
     }
 
+    void OnEnable()
+    {
+        ChangeGameState(GameState.Play);
+    }
+
+    public void ChangeGameState(GameState gameState)
+    {
+        _currentGameState = gameState;
+        OnGameStateChanged?.Invoke(_currentGameState);
+        Debug.Log((_currentGameState));
+    }
+
     public void IncreaseTotalEggCount()
     {
         _totalCollectedEggCount += 1;
@@ -30,9 +44,11 @@ public class GameManager : MonoBehaviour
         if (_totalCollectedEggCount == maxEggCount)
         {
             Debug.Log("Game Win!");
+            ChangeGameState(GameState.GameOver);
         }
     }
 
     public int TotalCollectedEggCount => _totalCollectedEggCount;
     public int MaxEggCount => maxEggCount;
+    public GameState CurrentGameState => _currentGameState;
 }